In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -12,24 +12,24 @@
 
 int main(void)
 {
-	int pro1, pro2 = 2, pro3 = 1, i = 1;
+	long pro1, pro2 = 2, pro3 = 1, i = 1;
 
 	printf("1, ");
 	printf("2, ");
 	while (i <= 16)
 	{
 		pro1 = pro2 + pro3;
-		printf("%u, ", pro1);
+		printf("%ld, ", pro1);
 		pro3 = pro1 + pro2;
-		printf("%u, ", pro3);
+		printf("%ld, ", pro3);
 		pro2 = pro3 + pro1;
 		if (i != 16)
 		{
-			printf("%u, ", pro2);
+			printf("%ld, ", pro2);
 		}
 		else
 		{
-			printf("%u\n, ", pro2);
+			printf("%ld\n, ", pro2);
 		}
 		i++;
 	}
